Move settings to sidenav and limit input widths
